etlo/requirements.txt
django==3.2.3
djoser==2.1.0
pillow==9.3.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
django_rest_passwordreset==1.3.0
django-oauth-toolkit==2.2.0
django_smtp_ssl==1.0
django-soft-delete==0.9.21
django-oauth2-provider==0.2.6.1
drf_social_oauth2==1.2.1
psycopg2-binary==2.8.6
selenium==4.7.2
beautifulsoup4==4.11.1
django-environ==0.9.0
django_smtp_ssl==1.0
django-storages==1.13.2
jsonfield==3.1.0
python-dateutil==2.8.2